pyproject.toml
[project]
name = "ida-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "anthropic>=0.64.0",
    "asyncpg>=0.30.0",
    "azure-identity>=1.24.0",
    "azure-storage-file-datalake>=12.21.0",
    "chainlit>=2.7.2",
    "google-adk>=1.13.0",
    "google-genai>=1.31.0",
    "greenlet>=3.2.4",
    "mcp>=1.13.1",
    "openai>=1.101.0",
    "openai-agents>=0.2.9",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "sqlalchemy>=2.0.43",
]
